requirements.txt
Flask
python-dotenv
requests
sounddevice
soundfile
openai 